Added utils.h
