+++ edge.conf
@@ -6,5 +6,4 @@ mac=DE:AD:BE:EF:10:01
 community=vde
 snodeport=1234
 snodaddr=localhost
-# WARNING!!
-# DO NOT modify the arrangement of the first 8 lines of this file!
+# WARNING!! DO NOT modify the arrangement of the above lines!
