+++ supernode.conf
@@ -1,1 +1,3 @@
--p=1234
+lport=1234
+daemon=0
+# WARNING!! DO NOT modify the arrangement of the above lines!
